entries and readings functionality changed for private entry filtering and fixing reading list ids
