+++ app/data/auth.jsx
@@ -25,7 +25,7 @@ export function getUserProfile() {
   const token = localStorage.getItem("token");
   if (!token) return Promise.resolve(null);
 
-  return fetchWithResponse("profile", {
+  return fetchWithResponse("profile/", {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
