+++ app/entries/entriesList.tsx
@@ -3,6 +3,7 @@
 import React, { useEffect, useState } from "react";
 import { EntryCard } from "../components/entryCard";
 import { getEntries } from "../data/entries";
+import { getUserProfile } from "../data/auth";
 
 interface User {
   id: number;
@@ -29,19 +30,25 @@ export default function EntriesList() {
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    async function fetchEntries() {
-      try {
-        const data = await getEntries();
-        const publicEntries = data.filter((entry) => !entry.isPrivate)
-        setEntries(publicEntries);
-      } catch (error) {
-        console.error("Error loading entries:", error);
-      } finally {
-        setLoading(false);
-      }
+  async function fetchEntries() {
+    try {
+      const profile = await getUserProfile()
+      const data = await getEntries();
+
+      const visibleEntries = data.filter(
+        (entry) => !entry.isPrivate || entry.user?.id === profile.id
+      );
+
+      setEntries(visibleEntries);
+    } catch (error) {
+      console.error("Error loading entries:", error);
+    } finally {
+      setLoading(false);
     }
-    fetchEntries();
-  }, []);
+  }
+  fetchEntries();
+}, []);
+
 
   if (loading) return <p>Loading entries...</p>;
