+++ app/entries/myentries/new/page.tsx
@@ -1,5 +1,16 @@
-import EntryForm from "@/app/components/EntryForm";
+'use client'
 
-export default function NewEntry() {
-  return <EntryForm />;
+import { useRouter } from "next/navigation";
+import EntryForm from "@/app/components/EntryForm";
+import { addEntry } from "@/app/data/entries";
+
+export default function CreateEntryPage() {
+  const router = useRouter();
+
+  const handleCreate = async (data: unknown) => {
+    await addEntry(data);
+    router.push("/entries/myentries");
+  };
+
+  return <EntryForm onSubmit={handleCreate} submitLabel="Create Entry" />;
 }
